+++ lib/providers/mistake_provider.g.dart
@@ -6,12 +6,12 @@ part of 'mistake_provider.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$mistakeNotifierHash() => r'fc44987d23649ecbb7376c3b067f2bcda45b639a';
+String _$mistakeNotifierHash() => r'ab66463c74c571c1c0dd156a9d06b8316fbbdaf0';
 
 /// See also [MistakeNotifier].
 @ProviderFor(MistakeNotifier)
 final mistakeNotifierProvider = AutoDisposeAsyncNotifierProvider<
-    MistakeNotifier, List<LessonBlock>>.internal(
+    MistakeNotifier, List<SentencePair>>.internal(
   MistakeNotifier.new,
   name: r'mistakeNotifierProvider',
   debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
@@ -21,6 +21,6 @@ final mistakeNotifierProvider = AutoDisposeAsyncNotifierProvider<
   allTransitiveDependencies: null,
 );
 
-typedef _$MistakeNotifier = AutoDisposeAsyncNotifier<List<LessonBlock>>;
+typedef _$MistakeNotifier = AutoDisposeAsyncNotifier<List<SentencePair>>;
 // ignore_for_file: type=lint
 // ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package